backend/requirements.txt
Flask
flask-cors
requests
sentence-transformers
python-dotenv
elasticsearch
